refactor(windows): u16 を VkCode に統一し VK/scancode の取り違えを型で防止
これまで `RawKeyEvent.vk_code: VkCode` という型境界の薄皮だけで、
awase-windows 内部の routing/state/output/IME パイプラインは
すべて u16 で実装されていた。NO_VK_AS_SCAN lint が守れていた範囲も
その薄皮の周辺だけで、HashSet/フィールド/関数引数に u16 で逃がした瞬間
保護外になっていた。

本コミットで Windows API 境界 (VIRTUAL_KEY, GetAsyncKeyState) 以外の
u16 を VkCode に揃え、コンパイラに取り違えを阻止させる。

主な変更:
- state/hook_state: ビットセット API (mark_engine_sent 等 8 関数) と
  HookConfig フィールドを VkCode 化
- hook: classify_route, classify_keyup_route, is_shortcut_bypass,
  is_non_shift_modifier, vk_raw フィールドを VkCode 化
- output: VkSequence, ascii_to_vk, symbol_to_vk, Vk(...) enum,
  defer_vk_if_probe_in_flight を VkCode 化
- output/vk_send, output/probe_io, tsf/output, tsf/probe_fsm,
  tsf/cold_warmup, tsf/send: send pipeline 全体を VkCode 化
- ime: send_ime_mode_key, push_release/restore 等を VkCode 化。
  ローカル const VK_KANJI/VK_RETURN/VK_F13/VK_F14 を crate::vk に置換
- panic_detect: PanicTriggerCombo.vk と is_panic_trigger を VkCode 化
- executor: deferred_passthrough_vks: HashSet<VkCode> 化
- platform: engine_on/off_ime_vk: Option<VkCode> 化
- runtime: MODIFIER_VKS テーブルを [VkCode; 11] 化

API 境界の残った `.0` は VIRTUAL_KEY(vk.0), i32::from(vk.0),
ビットセットの (vk.0 as usize) のみに局所化。

+++ crates/awase-windows/src/app/mod.rs
@@ -482,7 +482,7 @@ fn reload_config() {
             .iter()
             .chain(ime_off.iter())
             .map(|k| crate::panic_detect::PanicTriggerCombo {
-                vk: u16::from(k.vk),
+                vk: k.vk,
                 ctrl: k.ctrl,
                 shift: k.shift,
                 alt: k.alt,
